ai_breach_watcher/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
elasticsearch[async]>=8.17.0
agno>=1.0.0
anthropic>=0.42.0
gitpython>=3.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
websockets>=14.0
httpx>=0.28.0
